Switch/vj_switch_requirements.txt
openstacksdk==0.101.0
